pyproject.toml
[project]
name = "k8s-slack-mcp"
version = "0.1.0"
description = "A Kubernetes-integrated MCP server with Slack bot functionality"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp>=1.9.2",
    "kubernetes>=32.0.1",
    "slack-bolt>=1.23.0",
    "python-dotenv>=1.1.0"
]

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project.scripts]
k8s-slack-mcp = "k8s_slack_mcp.__main__:main"
